+++ daemonctl/views.py
@@ -1,0 +1,2 @@
+def websocket_handler(request):
+    pass
