--- web-daemon-control/html/answer.html
@@ -1,8 +1,0 @@
-<html>
-    <head>
-        <meta http-equiv="refresh" content="1;url=/">
-    </head>
-    <body>
-        <p>{{ answer }}</p>
-    </body>
-</html>
